Update README.md
Weather-Journal App Project
Overview
This project requires you to create an asynchronous web app that uses Web API and user data to dynamically update the UI with weather data and their provided response in text field. Also, it is providing the current date.

Project Environment setup
Node and Express Environment
In this project, we are using Node and Express environments. Node and Express should be installed on the local machine. The project file server.js should require express() and should create an instance of their app using express.

The Express app instance should be pointed to the project folder with .html, .css, and .js files.

Project Dependencies
The ‘cors’ package should be installed in the project from the command line, required in the project file server.js, and the instance of the app should be setup to use cors().

The body-parser package should be installed and included in the project.

Local Server
The local server should be running and producing feedback to the Command-Line through a working callback function.

API Credentials
Create API credentials on OpenWeatherMap.com or use the existing API key inside.